feat: 카드 디테일 작업
