+++ src/app/pokemon-card/[code]/components/CardDetailContainer.jsx
@@ -1,0 +1,13 @@
+"use client";
+
+import CardInfo from "./CardInfo";
+import CardTradeList from "./CardTradeList";
+
+export default function CardDetailContainer({ data, tradeList }) {
+  return (
+    <>
+      <CardInfo data={data} />
+      <CardTradeList tradeList={tradeList} innerClassName="mt-8" />
+    </>
+  );
+}
